DEPRECATION: primary_home variable is gone
Before I let users define both primary_home and primary_user but it was
left a pointless field. I was solvig a theoretical use-case problem with
the real life problem of additional complexity.

Now the concept of primary_home is gone and replaced with the explicit
use of `/home/{{ primary_user }}`. One less variable to get between you
and your configuration. If a server needs a custom user that new
primary_user can be defined.